Add possibility to save and load selected sounds
To make it easier to reuse certain sounds, it's now
possible to save and load presets for each sound (correct,
wrong, player 1, player 2 etc.) Same simple way of
saving as the questions list, in a json array.
As the sounds are stored in a certain order, meaning
that all sounds needs a valid file path when saving and
loading.